added font shader changed pointer code-style

+++ include/engine/shaders/shader_program.cpp
@@ -7,11 +7,11 @@ using engine::ShaderProgram;
 using engine::Shader;
 
 
-ShaderProgram::ShaderProgram(const std::initializer_list<Shader *> &shaders)
+ShaderProgram::ShaderProgram(const std::initializer_list<Shader*>& shaders)
 {
 	m_shader_program_id = glCreateProgram();
 	m_shader_program_created = true;
-	for (auto &shader : shaders)
+	for (auto& shader : shaders)
 		glAttachShader(m_shader_program_id, shader->m_shader_id);
 	glLinkProgram(m_shader_program_id);
 
@@ -23,7 +23,7 @@ ShaderProgram::ShaderProgram(const std::initializer_list<Shader *> &shaders)
 		std::string info_log;
 		if (info_log_length > 0)
 		{
-			info_log.resize(info_log_length);
+			info_log.resize((unsigned)info_log_length);
 			glGetProgramInfoLog(m_shader_program_id, info_log_length, nullptr, &info_log[0]);
 		}
 
@@ -34,7 +34,7 @@ ShaderProgram::ShaderProgram(const std::initializer_list<Shader *> &shaders)
 		}
 		Config::getInstance().log(info_log, Config::WARNING);
 	}
-	for (auto &shader : shaders)
+	for (auto& shader : shaders)
 		glDetachShader(m_shader_program_id, shader->m_shader_id);
 }
 
@@ -54,28 +54,28 @@ void ShaderProgram::unbind() const
 	glUseProgram(0);
 }
 
-void ShaderProgram::setUniformBool(const std::string &uniform_name, const bool &value) const
+void ShaderProgram::setUniformBool(const std::string& uniform_name, const bool& value) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
 	glUniform1i(location, value);
 }
 
-void ShaderProgram::setUniformInt(const std::string &uniform_name, const int &value) const
+void ShaderProgram::setUniformInt(const std::string& uniform_name, const int& value) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
 	glUniform1i(location, value);
 }
 
-void ShaderProgram::setUniformUInt(const std::string &uniform_name, const unsigned &value) const
+void ShaderProgram::setUniformUInt(const std::string& uniform_name, const unsigned& value) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
 	glUniform1ui(location, value);
 }
 
-void ShaderProgram::setUniformFloat(const std::string &uniform_name, const float &value) const
+void ShaderProgram::setUniformFloat(const std::string& uniform_name, const float& value) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
@@ -83,7 +83,7 @@ void ShaderProgram::setUniformFloat(const std::string &uniform_name, const float
 	glUniform1f(location, value);
 }
 
-void ShaderProgram::setUniformVector3(const std::string &uniform_name, const glm::vec3 &vector) const
+void ShaderProgram::setUniformVector3(const std::string& uniform_name, const glm::vec3& vector) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
@@ -91,7 +91,7 @@ void ShaderProgram::setUniformVector3(const std::string &uniform_name, const glm
 	glUniform3fv(location, 1, glm::value_ptr(vector));
 }
 
-void ShaderProgram::setUniformMatrix3(const std::string &uniform_name, const glm::mat3 &matrix) const
+void ShaderProgram::setUniformMatrix3(const std::string& uniform_name, const glm::mat3& matrix) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
@@ -99,7 +99,7 @@ void ShaderProgram::setUniformMatrix3(const std::string &uniform_name, const glm
 	glUniformMatrix3fv(location, 1, GL_FALSE, glm::value_ptr(matrix));
 }
 
-void ShaderProgram::setUniformMatrix4(const std::string &uniform_name, const glm::mat4 &matrix) const
+void ShaderProgram::setUniformMatrix4(const std::string& uniform_name, const glm::mat4& matrix) const
 {
 	bind();
 	auto location = glGetUniformLocation(m_shader_program_id, uniform_name.c_str());
